Refactor Footer and Footer Styles
